+++ WebApp_ControleDeGastos/Repository/Interface/ICategory.cs
@@ -7,10 +7,10 @@ namespace WebApp_ControleDeGastos.Repository.Interface
     public interface ICategory
     {
         List<Category> GetAllCategory();
-        Category GetCategoryById(int id);
+        Category GetCategoryById(long id);
         Task<Category> AddCategory(Category category);
         Task<Category> UpdateCategory(Category category);
-        Task<bool> DeleteCategory(int id);
+        Task<bool> DeleteCategory(long id);
         Task<List<Category>> GetCategoryByName(string name);
     }
 }
