+++ WebApp_ControleDeGastos/Repository/Interface/IUser.cs
@@ -7,9 +7,9 @@ namespace WebApp_ControleDeGastos.Repository.Interface
     public interface IUser
     {
         Task<List<User>> GetAllUser();
-        Task<User> GetUserById(int id);
+        Task<User> GetUserById(long id);
         Task<User> AddUser(User user);
         Task<User> UpdateUser(User user);
-        Task<bool> DeleteUser(int id);
+        Task<bool> DeleteUser(long id);
     }
 }
